% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_G_Caracterizacion_SF.R
\docType{data}
\name{D_G_Caracterizacion_SF}
\alias{D_G_Caracterizacion_SF}
\title{Datos Generales Caracterización SF}
\format{
Un marco de datos de tipo SF con 13 columnas y filas variables.
}
\source{
Local
}
\usage{
D_G_Caracterizacion_SF(url, Dptos_Evaluar, Mpios_Evaluar)
}
\arguments{
\item{url}{Ruta al repositorio en donde se encuentra el informe referente a los datos
generales de caracterización.}

\item{Dptos_Evaluar}{Vector con los nombres de los departamentos objeto de monitoreo.}

\item{Mpios_Evaluar}{Vector con los nombres de los municipios objeto de monitoreo.}
}
\description{
El informe de caracterización es transformado a un archivo con Características
Espaciales - SF. Lo anterior con el objeto de llevar a cabo revisiones relacionadas
con la distancia entre las coordenadas sistematizadas y las pre-registradas.
Es importante comentar que el conjunto de datos se encuentra filtrado de acuerdo con los
departamentos y municipios objeto de monitoreo.
}
\author{
Mendoza-Ureche R. \email{romu545@gmail.com}
}
\keyword{D_G_Caracterizacion_SF}
