% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_G_Preregistro.R
\docType{data}
\name{D_G_Preregistro}
\alias{D_G_Preregistro}
\title{Datos Generales Pre-registro}
\format{
Un marco de XX columnas y filas variables.
}
\source{
Local
}
\usage{
D_G_Preregistro(url)
}
\arguments{
\item{url}{Ruta al repositorio en donde se encuentra el informe referente a los datos
generales del preregitro.}
}
\description{
Datos Generales Pre-registro
}
\author{
Mendoza-Ureche R. \email{romu545@gmail.com}
}
\keyword{D_G_Preregistro}
