% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos_Evaluar_SF.R
\name{Datos_Evaluar_SF}
\alias{Datos_Evaluar_SF}
\title{Datos Pre-registro Evaluar SF}
\usage{
Datos_Evaluar_SF(Datos.Evaluar)
}
\arguments{
\item{Datos.Evaluar}{Conjunto de datos para evaluar.}
}
\value{
Retorna el conjunto de datos de los pre-registros evaluados con características
espaciales.
}
\description{
Esta función transforma el conjunto de datos de los pre-registros para
evaluar a conjunto con características espaciales.
}
