% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exportar_Encuestas_Para_Habilitar.R
\name{Exportar_Encuestas_Para_Habilitar}
\alias{Exportar_Encuestas_Para_Habilitar}
\title{Exportar Encuestas Para Habilitar}
\usage{
Exportar_Encuestas_Para_Habilitar(
  Lista_Observaciones_x_Colector,
  Vector_Para_Filtro
)
}
\arguments{
\item{Lista_Observaciones_x_Colector}{Lista con las observaciones de cada colector.}

\item{Vector_Para_Filtro}{Vector booleano para filtrar los pre-registros que no posean
observaciones
param Semana Cadena de texto indicando la semana o semanas evaluadas.
  * `Semana_1`
  * `Semanas_5_6`
param Carpeta Nombre de la carpeta donde se exportaran los archivos que se generen}
}
\value{
Retorna un archivo .xlsx con los pre-registro que se habilitaran en primera instancia.
}
\description{
Esta función permite exportar aquellos pre-registros que no tengan alguna
observación
}
