% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exportar_Observaciones_x_Colector.R
\name{Exportar_Observaciones_x_Colector}
\alias{Exportar_Observaciones_x_Colector}
\title{Exportar Observaciones Pre-regsitros Por Colector}
\usage{
Exportar_Observaciones_x_Colector(
  Lista.Observaciones.x.Colector,
  Vector.Para.Filtro
)
}
\arguments{
\item{Lista.Observaciones.x.Colector}{Lista con las observaciones de cada colector.}

\item{Vector.Para.Filtro}{Vector booleano para filtrar los pre-registros que si posean
observaciones
param Semana Cadena de texto indicando la semana o semanas evaluadas.
  * `Semana_1`
  * `Semanas_5_6`
param Carpeta Nombre de la carpeta donde se exportaran los archivos que se generen}
}
\value{
Retorna un archivo .xlsx con los pre-registro que se habilitaran en primera instancia.
}
\description{
Esta función permite exportar aquellos pre-registros que no tengan alguna
observación
}
