% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obs_Altitudes_UPA.R
\name{Obs_Altitudes_UPA}
\alias{Obs_Altitudes_UPA}
\title{Observaciones Altitudes UPA}
\usage{
Obs_Altitudes_UPA(Datos.Elevacion)
}
\arguments{
\item{Datos.Elevacion}{Conjunto de datos para evaluar con las elevaciones calculadas.}
}
\value{
Retorna una tabla con el número de registro y sus respectivas observaciones.
}
\description{
Esta función permite generar observaciones cuando las altitudes registradas
quedan por fuera del intervalo generado mediante el modelo digital de elevación de AWS.
}
