% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obs_Coordenadas_En_Pregistros_Antiguos.R
\name{Obs_Coordenadas_En_Pregistros_Antiguos}
\alias{Obs_Coordenadas_En_Pregistros_Antiguos}
\title{Observaciones Coordenadas En Pre-registros Antiguos}
\usage{
Obs_Coordenadas_En_Pregistros_Antiguos(Datos.Evaluar, Preregistros.Antiguos)
}
\arguments{
\item{Datos.Evaluar}{Conjunto de datos para evaluar.}

\item{Preregistros.Antiguos}{Conjusto de datos de los pre-registros evaluados en cortes
anteriores}
}
\value{
Retorna una tabla con el número de registro y sus respectivas observaciones.
}
\description{
Esta función permite generar observaciones cuando las coordenadas de los
pre-registros evaluados estén digitadas en pre-registros evaluados en cortes anteriormente.
}
