% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obs_Coordenadas_Preregistro_vs_Asignadas.R
\name{Obs_Coordenadas_Preregistro_vs_Asignadas}
\alias{Obs_Coordenadas_Preregistro_vs_Asignadas}
\title{Observaciones UPA Pre-registro vs Asignadas}
\usage{
Obs_Coordenadas_Preregistro_vs_Asignadas(Datos_Evaluar_SF, UPA_Asignadas_SF)
}
\arguments{
\item{Datos_Evaluar_SF}{Conjunto de datos para evaluar con características espaciales.}

\item{UPA_Asignadas_SF}{Conjunto de datos de UPA asignadas con características espaciales.}
}
\value{
Retorna una tabla con el número de registro y sus respectivas observaciones.
}
\description{
Esta función permite generar observaciones cuando se encuentra alguna
inconsistencia al constrastar las UPA pre-registradas con las asignadas.
}
