% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obs_Distancia_Con_UPA_Digitadas.R
\name{Obs_Distancia_Con_UPA_Digitadas}
\alias{Obs_Distancia_Con_UPA_Digitadas}
\title{Observaciones Distancia Entre UPA Pre-registradas y Digitadas}
\usage{
Obs_Distancia_Con_UPA_Digitadas(Datos_Evaluar_SF, Datos_Caracterizacion_SF)
}
\arguments{
\item{Datos_Evaluar_SF}{Conjunto de datos para evaluar con características espaciales.}

\item{Datos_Caracterizacion_SF}{Datos caracterización con características espaciales}
}
\value{
Retorna una tabla con el número de registro y sus respectivas observaciones.
}
\description{
Esta función permite generar observaciones cuando las coordenadas de los
pre-registros evaluados se encuentran cerca de coordenadas registradas en el formulario
de caracterición.
}
