% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obs_Municipios_Diferentes.R
\name{Obs_Municipios_Diferentes}
\alias{Obs_Municipios_Diferentes}
\title{Observaciones Municipio Distinto}
\usage{
Obs_Municipios_Diferentes(Datos.Evaluar.SF, Mpio.shp)
}
\arguments{
\item{Datos.Evaluar.SF}{Conjunto de datos para evaluar con características espaciales.}

\item{Mpio.shp}{Capa de municipios}
}
\value{
Retorna una tabla con el número de registro y sus respectivas observaciones.
}
\description{
Esta función permite generar observaciones cuando las coordenadas ubican
las coordenadas pre-registras en un municipio distinto al seleccionado.
}
