% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obs_Nombre_Predio.R
\name{Obs_Nombre_Predio}
\alias{Obs_Nombre_Predio}
\title{Observaciones Nombre Predio}
\usage{
Obs_Nombre_Predio(Datos_Evaluar, Nom_Evaluar)
}
\arguments{
\item{Datos_Evaluar}{Conjunto de datos para evaluar.}

\item{Nom_Evaluar}{Vector con nombres que se tendrán en cuenta en la revisión.}
}
\value{
Retorna una tabla con el número de registro y sus respectivas observaciones.
}
\description{
Esta función evalúa los nombres de los predios en donde no se evidenció
actividad acuícola con el objeto de verificar que estos sean diligenciados con algún
nombre en especifico.
}
