% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obs_Nombre_UPA_Inexistente.R
\name{Obs_Nombre_UPA_Inexistente}
\alias{Obs_Nombre_UPA_Inexistente}
\title{Observaciones Nombre UPA Inexistente}
\usage{
Obs_Nombre_UPA_Inexistente(Datos.Evaluar, Nom.No.Evaluar)
}
\arguments{
\item{Datos.Evaluar}{Conjunto de datos para evaluar.}

\item{Nom.No.Evaluar}{Vector con nombres que no se tendrán en cuenta en la revisión.}
}
\value{
Retorna una tabla con el número de registro y sus respectivas observaciones.
}
\description{
Esta función evalúa los nombres de los sitios en donde no se evidenció
activida acuicola con el objeto de verificar que estos sean diligenciado como
La UPA no existe.
}
