% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Relacion_General_Observaciones.R
\name{Relacion_General_Observaciones}
\alias{Relacion_General_Observaciones}
\title{Consolidado General Observaciones}
\usage{
Relacion_General_Observaciones(Datos_Evaluar)
}
\arguments{
\item{Datos_Evaluar}{Conjunto de datos para evaluar}
}
\value{
Retorna una tabla con todos los registros evaluados y sus respectivas observaciones
}
\description{
Esta función realiza el consolidado de todas las observaciones generadas.
}
