% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vector_Para_Filtro.R
\name{Vector_Para_Filtro}
\alias{Vector_Para_Filtro}
\title{Vector Booleano Para Filtro}
\usage{
Vector_Para_Filtro(Lista_Observaciones_x_Colector)
}
\arguments{
\item{Lista_Observaciones_x_Colector}{Lista de observaciones de cada colector.}
}
\value{
Retorna un vector booleano
}
\description{
Esta función permite crear un vector booleano que filtrar en primera
instancia los pre-registros sin observaciones que se enviaran a habilitar de los
que poseen alguna observación para ser validada por los colectores.
}
