% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos_Elevacion.R
\name{Datos_Elevacion}
\alias{Datos_Elevacion}
\title{Datos Elevaciones Datos Evaluados}
\usage{
Datos_Elevacion(Datos_Evaluar_SF)
}
\arguments{
\item{Datos_Evaluar_SF}{Conjunto de datos para evaluar con características espaciales.}
}
\value{
Retorna una tabla con el número de registro y las altitudes registradas y calculadas.
}
\description{
Esta función permite obtener la altitud de las coordenadas evaluadas
mediante por fuera el modelo digital de elevación de AWS.
}
