% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lista_Observaciones_x_Colector.R
\name{Lista_Observaciones_x_Colector}
\alias{Lista_Observaciones_x_Colector}
\title{Lista Observaciones Por Colector}
\usage{
Lista_Observaciones_x_Colector(Relacion_General_Observaciones)
}
\arguments{
\item{Relacion_General_Observaciones}{Conjunto de datos del consolidado general de observaciones}
}
\value{
Retorna una lista con las observaciones por cada colector
}
\description{
Esta función crea una lista con las observaciones por cada colector.
}
