% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obs_Codigo_Misma_UPA.R,
%   R/Obs_Coordenadas_Repetidas.R
\name{Obs_Codigo_Misma_UPA}
\alias{Obs_Codigo_Misma_UPA}
\alias{Obs_Coordenadas_Repetidas}
\title{Observaciones Otros Códigos Misma UPA}
\usage{
Obs_Codigo_Misma_UPA(Datos.Evaluar, Datos.Preregistro)

Obs_Coordenadas_Repetidas(Datos.Evaluar)
}
\arguments{
\item{Datos.Evaluar}{Conjunto de datos para evaluar.}

\item{Datos.Preregistro}{Conjunto de datos de todos los pre-registros.}
}
\value{
Retorna una tabla con el número de registro y sus respectivas observaciones.

Retorna una tabla con el número de registro y sus respectivas observaciones.
}
\description{
Esta función evalúa que los códigos registrados en el campo para otros id
para una misma UPA no hayan sido utilizados en algún pre-registro como el identificador
principal.

Esta función permite generar observaciones cuando existan más de dos
pre-registros con las mismas coordenadas.
}
