% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obs_Nombre_UPA.R
\name{Obs_Nombre_UPA}
\alias{Obs_Nombre_UPA}
\title{Observaciones Nombre UPA}
\usage{
Obs_Nombre_UPA(Datos_Evaluar, Lista_Referencia_Nombres, Nom_No_Evaluar)
}
\arguments{
\item{Datos_Evaluar}{Conjunto de datos para evaluar.}

\item{Lista_Referencia_Nombres}{Listado de nombres catalogados como validados.}

\item{Nom_No_Evaluar}{Vector con nombres que no se tendrán en cuenta en la revisión.}
}
\value{
Retorna una tabla con el número de registro y sus respectivas observaciones.
}
\description{
Esta función evalúa los nombres de las unidades productivas con el objeto
de verificar que estos estén escritos correctamente.
}
