% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obs_UPA_Repetidas.R
\name{Obs_UPA_Repetidas}
\alias{Obs_UPA_Repetidas}
\title{Observaciones UPA Repetidas}
\usage{
Obs_UPA_Repetidas(Datos.Evaluar, Nom.No.Evaluar)
}
\arguments{
\item{Datos.Evaluar}{Conjunto de datos para evaluar.}

\item{Nom.No.Evaluar}{Vector con nombres que no se tendrán en cuenta en la revisión.}
}
\value{
Retorna una tabla con el número de registro y sus respectivas observaciones.
}
\description{
Esta función permite generar observaciones cuando los valores en los campos
Nombre de la UPA, Departamento, Municipio, Centro poblado, Latitud y Longitud aparezcan
más de un pre-registro.
}
